requirements.txt
selenium==4.15.2
undetected-chromedriver==3.5.4
gspread==5.12.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-auth==2.25.2
python-dotenv==1.0.0
pyotp==2.9.0
imapclient==3.0.1
pillow==10.1.0
requests==2.31.0
selenium-stealth==1.0.1
